refactor: satisfiesPkg back to two-state; break all module cycles
- The 'default' package-format query only exists on the download
  resolution path (ReleaseService filters / assetMatchesPlatform);
  nothing that calls the structural PecansAssetQuery ever passes it, so
  satisfiesPkg returns to plain 'pkg == undefined || this.pkg == pkg'
  and PackageFormatQuery moves next to platformToQuery in
  utils/platforms.
- All modules now import specific util/model modules instead of the
  barrels, dissolving every import cycle (the barrels re-enter service
  through the deprecated resolveForVersion adapter); madge --circular
  reports none (Copilot review x2).
