+++ src/backends/github.ts
@@ -11,7 +11,11 @@ import {
   PecansReleaseDTO,
   isPecansAsset,
 } from "../models";
-import { channelFromVersion, filenameToPlatform } from "../utils/";
+// specific module imports to stay out of the ../utils barrel - it
+// re-exports the deprecated resolveForVersion adapter, which imports the
+// service module, which imports this backend (import cycle)
+import { channelFromVersion } from "../utils/channelFromVersion";
+import { filenameToPlatform } from "../utils/platforms";
 import { PecansReleases } from "../models/PecansReleases";
 import { clean } from "semver";
