+++ src/models/PecansAssetQuery.ts
@@ -1,19 +1,21 @@
-import { Architecture, OperatingSystem, PackageFormat } from "../utils";
+// import from the specific util modules, not the ../utils barrel - the
+// barrel re-exports resolveForVersion, whose deprecation adapter imports
+// the service module that consumes these query types (import cycle)
+import { Architecture } from "../utils/Architecture";
+import { OperatingSystem } from "../utils/OperatingSystem";
+import { PackageFormat } from "../utils/PackageFormat";
 import { SupportedFileExtension } from "../utils/SupportedFileExtension";
-
-/**
- * Package-format criterion. An asset's pkg is undefined for the platform's
- * default package (tarball, dmg, setup.exe) and set for alternate formats
- * (deb, rpm). Queries are unconstrained when pkg is omitted; "default"
- * selects only default-package assets.
- */
-export type PackageFormatQuery = PackageFormat | "default";
 
 export interface PecansAssetQuery {
   os?: OperatingSystem;
   arch?: Architecture;
-  /** undefined = any package format; "default" = the platform default only */
-  pkg?: PackageFormatQuery;
+  /**
+   * Alternate package format to require (deb/rpm); omitted = unconstrained.
+   * An asset without an alternate format is the platform's default package
+   * (tarball, dmg, setup.exe) - selecting only those is a download-semantics
+   * concern, expressed as pkg: "default" on the ReleaseService filters.
+   */
+  pkg?: PackageFormat;
   filename?: string;
   extensions?: SupportedFileExtension[];
 }
