+++ src/utils/platforms.ts
@@ -119,13 +119,21 @@ export function parsePlatform(platform: Platform): PlatformParts {
   };
 }
 
+/**
+ * Package-format criterion for download resolution. An asset without an
+ * alternate package format (deb/rpm) IS the platform's default package
+ * (tarball, dmg, setup.exe); "default" selects exactly those. Omitted =
+ * unconstrained.
+ */
+export type PackageFormatQuery = PackageFormat | "default";
+
 /** Discrete query equivalent of a composite platform id. */
 export interface DiscretePlatformQuery {
   os: OperatingSystem;
   /** undefined = any architecture */
   arch?: Architecture;
   /** undefined = any package format; "default" = the platform default only */
-  pkg?: PackageFormat | "default";
+  pkg?: PackageFormatQuery;
 }
 
 /**
